+++ RayTracing/src/Hittable/sphere.hpp
@@ -4,16 +4,15 @@
 
 class Sphere : public Hittable {
 public:
-    Sphere(const Point& center, double radius, std::shared_ptr<const Material> material);
-
+    Sphere(const Point& center, float radius, std::shared_ptr<const Material> material);
     const Point& center() const;
-    double radius() const;
-    std::optional<HitRecord> hit(const Ray& ray, double t_min, double t_max) const override;
-    std::optional<double> get_t_if_hit(const Ray& ray, double t_min, double t_max) const override;
-    HitRecord get_hit_results(const Ray& ray, double t) const override;
+    float radius() const;
+    std::optional<HitRecord> hit(const Ray& ray, float t_min, float t_max) const override;
+    std::optional<float> get_t_if_hit(const Ray& ray, float t_min, float t_max) const override;
+    HitRecord get_hit_results(const Ray& ray, float t) const override;
 
 private:
     const Point C;
-    const double r;
+    const float r;
     std::shared_ptr<const Material> material;
 };
